+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add support for TYPO3 10LTS (#1136)
 
 ### Changed
+- Do not advertise the crowdfunding campaign (#1150)
 - Allow installations with Emogrifier 5.x and 6.x (#1147)
 - Require oelib >= 4.1 (#1143, #1146)